requirements.txt
pywin32>=306; sys_platform == "win32"
playwright>=1.40.0
openpyxl>=3.1.2
flask>=3.0.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
pyyaml>=6.0.1
msal>=1.28.0
requests>=2.31.0
